+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Release Versions
 
 - fix: raise exception when accessing an empty parameter value in python (#260)
 - feat: move up to C++20 (#220)
+- feat: split trajectory classes and implement Cartesian/JointState specializations (#216)
 
 ## 9.3.1